chore: rename documentationCache documentation-cache
