+++ src/utils/documentation-cache.ts
